+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Version 0.2.0 (2019-07-16)
 - FEAT: implement proper git <-> helper command exchange so that git no longer exits with non-0 code
 - ENH: factor out `cli` to `cli_core` and `Ec2Class` in preparation for new `aws+cw` protocol
 - FEAT: factor out and add alternative command for cloudwatch
+- FEAT: add cloudwatch/describe-alarms
 
 
 Version 0.1.0 (2019-07-12)